JiraId:ZXQ-2416:解决对不同车辆请求远程车控，返回命令ID错误的BUG

+++ sp-vp-service/src/main/java/com/zxq/iov/cloud/sp/vp/service/impl/TboxConfigServiceImpl.java
@@ -67,7 +67,7 @@ public class TboxConfigServiceImpl extends BaseService implements ITboxConfigSer
         if(null == tboxPersonalConfig) {
             throw new ServLayerException(ExceptionConstants.PERSONAL_CONFIG_NOT_FIND);
         }
-        if(tboxPersonalConfig.getTboxId().intValue() != tboxId) {
+        if(tboxPersonalConfig.getTboxId().intValue() != tboxId.intValue()) {
             throw new ServLayerException(ExceptionConstants.TBOX_NOT_MATCH_VIN);
         }
         if(tboxPersonalConfig.getConfigDelta().intValue() > configDelta.intValue()) {
